refactor: Remove services/
