+++ src/app/exchange-rate-info/converter/converter.component.ts
@@ -3,7 +3,6 @@ import { CommonModule, AsyncPipe } from '@angular/common';
 import { ExchangeRateInfo } from '../exchange-rate-infos.type';
 import { ExchangeRateCardComponent } from 'src/app/exchange-rate-info/exchange-rate-card/exchange-rate-card.component';
 import { ExchangeRateChartComponent } from 'src/app/exchange-rate/exchange-rate-chart.component';
-import { ExchangeRateService } from 'src/app/services';
 import { FormBuilder, FormsModule, ReactiveFormsModule, Validators } from '@angular/forms';
 import { GridDirective } from '../../grid.directive';
 import { LetDirective } from '@ngrx/component';
@@ -17,6 +16,7 @@ import { SubscriptionSupervisorComponent } from 'src/app/subscription-supervisor
 import { combineLatest, filter, interval, mergeMap, Observable, share, startWith, tap, withLatestFrom } from 'rxjs';
 import { ExchangeRate } from '../../exchange-rate/exchange-rate.type';
 import { iso4217 } from 'src/app/exchange-rate/iso4217.enum';
+import { ExchangeRateService } from '../exchange-rate.service';
 
 @Component({
     selector: 'app-converter',
